refactor: share forecast and report command helpers
The forecast and report commands duplicated the same refreshed-account
persistence flow and forecast-result serialization logic. That increases
maintenance cost and makes future fixes easy to miss in one path.

Extract the shared pieces into a single helper module and keep both
commands delegating to the same implementation.
